fix: 修复avg_volume类型为Decimal导致除法失败
将avg_volume转换为float避免与float类型的avg_window_volume运算时出错

+++ crypto-analyzer/app/services/daily_review_analyzer.py
@@ -290,7 +290,7 @@ class DailyReviewAnalyzer:
             return []
 
         # 计算平均成交量
-        avg_volume = sum(k['volume'] for k in klines) / len(klines)
+        avg_volume = float(sum(k['volume'] for k in klines) / len(klines))
 
         opportunities = []
         threshold = self.thresholds[timeframe]
